Fix and improvise constants.ts.
