+++ src/trxs/krc20/krc20Api.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosError } from 'axios';
 import axiosRetry from 'axios-retry';
 import Monitoring from '../../monitoring';
-import CONFIG from "../../../config/constants";
+import CONFIG from "../../constants";
 
 export let KASPLEX_URL = 'https://api.kasplex.org'
 if( CONFIG.network === "testnet-10" ) {
